Update assertOk to handle string and byte types more generally
The python3 porting changes introduced an assumption
that decode() can always be called on the second arg
for assertOk - this is not necessarily the case, so
it now transparently accepts either case of bytes vs
string.

Also add an assertNotExists interface to cluster.py,
as was done for the __init__.py case earlier.